refactor: build root screen with a Consumer
